rename: 커뮤니티 파일명 변경
