+++ src/app/community/article/article.module.ts
@@ -4,9 +4,9 @@ import { ArticleService } from './article.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ArticleLike } from '@domain/communities/articles/article-like.entity';
 import { Article } from '@domain/communities/articles/article.entity';
-import { BoardModule } from '@app/communities/board/board.module';
+import { BoardModule } from '@app/community/board/board.module';
 import { UserModule } from '@app/user/user.module';
-import { ArticlePhotoClient } from '@app/communities/article/utils/article-photo.client';
+import { ArticlePhotoClient } from '@app/community/article/utils/article-photo.client';
 
 @Module({
   imports: [
