+++ src/app/community/article/commands/article-delete.command.ts
