+++ src/app/community/article/commands/article-preview.command.ts
