+++ src/app/community/article/commands/article-profile.command.ts
