+++ src/app/community/article/dtos/article-create.request.ts
@@ -1,4 +1,4 @@
-import { ArticleCreateCommand } from '@app/communities/article/commands/article-create.command';
+import { ArticleCreateCommand } from '@app/community/article/commands/article-create.command';
 
 export class ArticleCreateRequest implements Partial<ArticleCreateCommand> {
   title: string;
