+++ src/app/community/article/dtos/article-update.request.ts
@@ -1,4 +1,4 @@
-import { ArticleUpdateCommand } from '@app/communities/article/commands/article-update.command';
+import { ArticleUpdateCommand } from '@app/community/article/commands/article-update.command';
 
 export class ArticleUpdateRequest implements Partial<ArticleUpdateCommand> {
   title: string;
