+++ src/app/community/article/utils/article-photo.client.ts
