+++ src/app/community/board/board.module.ts
