+++ src/app/community/board/commands/board-delete.command.ts
