+++ src/app/community/board/commands/board-profile.command.ts
