+++ src/app/community/comment/dtos/create-comment.request.ts
@@ -1,4 +1,4 @@
-import { CommentCreateCommand } from '@app/communities/comments/commands/comment-create.command';
+import { CommentCreateCommand } from '@app/community/comment/commands/comment-create.command';
 
 export class CreateCommentRequest implements Partial<CommentCreateCommand> {
   content: string;
